+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## master (unreleased)
 
+- Add ability to specify a field name explicitly. ([@palkan][])
+
+Now you can write, for example:
+
+```ruby
+expose_authorization_rules :create?, with: PostPolicy, field_name: :can_create_post
+```
+
 - Add support for resolvers. ([@palkan][])
 
 Now it's possible to `include ActionPolicy::GraphQL::Behaviour` into resolver class to use